nokia_site: add function to unlock and reset TRXes
We have a long outstanding issue with TRX RSL bootstrap,
espcially with BTSes with more than one TRX. The more TRX
a BTS has, the more time it takes for it get to a ready state.

This commit chnages the TRX bootstrap logic as the following:

1. During OML initial config, all TRXes are confgiured locked.
   This gives a better controled state of TRXes, and the BTS
   enough time to do internal checks, SW load etc.
2. BTS_CONF_COMPL as entry point for the RSL bootstrap.
3. Unlock all TRXes, and give the BCCH TRX a reset.
4. Start RSL bootstrap.

Runtime tested with InSite and a MetroSite with 4 TRXes.